+++ Repositories/Interfaces/IProductRepository.cs
@@ -9,7 +9,7 @@ namespace HadiyahRepositories.Interfaces
 {
     public interface IProductRepository : IRepository<Product>
     {
-        Task<IEnumerable<Product>> GetByCategoryAsync(int categoryId);
+        Task<IEnumerable<Product>> GetByCategoryAsync(long categoryId);
         Task<bool> ExistsByNameAsync(string name);
     }
 }
